third_party/WebKit/LayoutTests/xr/exclusive_requestFrame_called.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../external/wpt/resources/chromium/webxr-test.js"></script>
<script src="../xr/resources/xr-test-utils.js"></script>
<canvas id="webgl-canvas"></canvas>

<script>
let testName = "XRSession requestAnimationFrame properly calls the provided "
  + "callback";

let fakeDeviceInitParams = { supportsImmersive:true };

let requestSessionOptions = [
  { immersive: true },
  { outputContext: getOutputContext() }
];

let testFunction = (session) => new Promise((resolve) => {
  // Session must have a baseLayer or frame requests will be ignored.
  session.baseLayer = new XRWebGLLayer(session, gl);

  function onFrame(time, xrFrame) {
    assert_true(xrFrame instanceof XRFrame);
    // Test does not complete until the returned promise resolves.
    resolve();
  }

  session.requestAnimationFrame(onFrame);
});

xr_session_promise_test(
  testFunction, fakeDeviceInitParams, requestSessionOptions, testName);

</script>
